Split app into modules
